+++ src/main/java/com/web/chon/negocio/NegocioCatCorreos.java
@@ -1,5 +1,6 @@
 package com.web.chon.negocio;
 import com.web.chon.dominio.Correos;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import javax.ejb.Remote;
@@ -13,7 +14,7 @@ public interface NegocioCatCorreos
 {
     public int insertCorreo(Correos co);
     public int deleteCorreos(Correos co);
-    public List<Object[]> SearchCorreosbyidClientPk(int idClientepk);
+    public List<Object[]> SearchCorreosbyidClientPk(BigDecimal idClientepk);
     public int updateCorreos(Correos co);
     
 }
